[FEATURE ember-application-engines] Move initializers from Application to Engine.
Initializers and instance initializers are a common concern for
applications and engines.
